edited the main.c in test folder
